* m4/libelf.m4: Resync with version of code in aclocal.m4.
git-svn-id: svn://sourceware.org/svn/prelink/trunk@176 94c539fb-cf18-0410-b60f-edeeb537fa16

+++ m4/libelf.m4
@@ -73,7 +73,8 @@ int main (void)
   Elf64_Addr val;
   Elf_Scn *scn;
 
-  fd = open ("conflibelftest", O_RDWR | O_CREAT | O_EXCL); if (fd == -1) exit (1);
+  fd = open ("conflibelftest", O_RDWR | O_CREAT | O_EXCL, 0600);
+  if (fd == -1) exit (1);
   unlink ("conflibelftest");
   if (write (fd, sparc64_elf, sizeof(sparc64_elf)) != sizeof(sparc64_elf))
     exit (1);
